move the load utils so that they don't appear in ./mange.py help
